[MCH] adapted raw data decoder to UserLogic format
- moved CRU link mapping into channel handler code
- added simple UserLogic decoder that uses a hand-written FSM similar to that of the bare decoder
- removed the "FORMAT" template parameter, as the data  format is now deduced directly from the data
- improved detection of orbit jumps, by instantiating one PayloadParser for each CRU link

In order to avoid over-instantiation of objects, the PayloadParser objects are created dynamically

+++ Detectors/MUON/MCH/Raw/ElecMap/include/MCHRawElecMap/CruLinkId.h
@@ -32,8 +32,8 @@ class CruLinkId
   uint16_t mDeId;
 };
 
-CruLinkId decodeCruLinkId(uint32_t code);
-uint32_t encode(const CruLinkId& id);
+CruLinkId decodeCruLinkId(uint16_t code);
+uint16_t encode(const CruLinkId& id);
 
 std::ostream& operator<<(std::ostream& os, const CruLinkId& id);
